packages/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0

# Image processing
opencv-python-headless>=4.9.0
pillow>=10.2.0

# AI/ML
google-generativeai>=0.3.0

# Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
